Revert "Modify gateway "
